rename well transform and move utils into folder
