+++ packages/components/codemods/utils/index.ts
@@ -1,2 +1,4 @@
 export * from "./transformSource"
 export * from "./getImportAlias"
+export * from "./printAst"
+export * from "./transformComponentsInDir"
