+++ packages/components/codemods/utils/printAst.ts
@@ -1,5 +1,5 @@
 import ts from "typescript"
-// This function is used to print the AST as a string and is only used for testing
+/** A util that prints that is used to print the AST as a string */
 export const printAst = (ast: ts.SourceFile): string => {
   const printer = ts.createPrinter()
   return printer.printFile(ast)
